+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ CHANGELOG
 5.1.19 (unreleased)
 -------------------
 
+- Update default zope.interface to 4.7.1
+  [bloodbare]
+
 - Be able to provide `DEBUG_SUBSCRIBERS` env variable to get details about
   event timings being run.
   [vangheem]